Cleaned up code and styling (except for padding/margin formatting)
Cleaned up the code for the About Us page to reduce redundancy and make the code more uniform (and easier to find and fix). The only thing that hasn't been fixed yet is the padding/margin issues that make the design not line-up with the Figma design. This is due to code to the "Padding" and "Margin" styling used in the "content-container" class.